backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# Vector Store
chromadb>=0.4.22

# Document Processing
pymupdf>=1.23.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.0
